+++ src/db/companies.ts
@@ -1,0 +1,9 @@
+import { pgTable, text, uuid } from "drizzle-orm/pg-core";
+
+// --- Companies ( offerT14X) ---
+
+export const companies = pgTable('companies', {
+  id: uuid('id').primaryKey().defaultRandom(),
+  name: text('name').notNull(),
+  website: text('website'),
+});
