+++ src/db/schema.ts
@@ -2,5 +2,8 @@ import { account, session, user, verification } from "./auth-schema"
 import { tools } from "./tools";
 import { categories } from "./categories";
 import { platforms } from "./platforms";
+import { tags } from "./tags";
+import { companies } from "./companies";
+import { features } from "./features";
 
-export { account, session, user, verification, tools, categories, platforms }
+export { account, session, user, verification, tools, categories, platforms, tags, companies, features }
