+++ src/db/tags.ts
@@ -1,0 +1,9 @@
+import { pgTable, text, uuid } from "drizzle-orm/pg-core";
+
+// --- Tags ---
+
+
+export const tags = pgTable('tags', {
+  id: uuid('id').primaryKey().defaultRandom(),
+  name: text('name').notNull(),
+});
